update folder structure for single page app
