+++ src/index.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+
 import { ChakraProvider } from '@chakra-ui/react';
 import theme from './utils/theme';
