--- src/pages/resume/Resume.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Center } from '@chakra-ui/react';
-
-const Resume = () => {
-  return (
-    <Center className="resume" style={{ minHeight: 'calc(100vh - 205px)' }}>
-      I&apos;ll add this later...
-    </Center>
-  );
-};
-
-export default Resume;
